requirements.txt
oauthlib==3.1.1
PySocks==1.7.1
python-dotenv==0.18.0
requests-oauthlib==1.3.0
tweepy==3.10.0
